refactor: Move TokenIcon and TokenIcons to UI
