+++ apps/main/src/components/PageDashboard/index.tsx
@@ -10,7 +10,6 @@ import { REFRESH_INTERVAL } from '@/constants'
 import { breakpoints, shortenAccount } from '@/ui/utils'
 import { sleep } from '@/utils'
 import useStore from '@/store/useStore'
-import useTokensMapper from '@/hooks/useTokensMapper'
 
 import { ExternalLink } from '@/ui/Link'
 import Table from '@/ui/Table'
@@ -32,7 +31,6 @@ const Dashboard = ({
   const isSubscribed = useRef(false)
   const formValuesWalletAddressRef = useRef('')
 
-  const { tokensMapper } = useTokensMapper(rChainId)
   const activeKey = useStore((state) => state.dashboard.activeKey)
   const isValidWalletAddress = useStore((state) => state.dashboard.isValidWalletAddress[activeKey])
   const dashboardError = useStore((state) => state.dashboard.error)
@@ -176,7 +174,6 @@ const Dashboard = ({
                     poolData={poolData}
                     haveBoost={haveBoost}
                     tableLabel={TABLE_LABEL}
-                    tokensMapper={tokensMapper}
                     walletAddress={formValues.walletAddress}
                     walletPoolData={data}
                   />
