+++ packages/ui/src/Token/index.ts
@@ -1,0 +1,2 @@
+export { default as TokensIcon } from 'Token/TokensIcon'
+export { default } from './TokenIcon'
