+++ packages/ui/src/index.tsx
@@ -52,6 +52,7 @@ export * from './TableButtonFiltersMobile'
 export * from './TableSort'
 export * from './TextCaption'
 export * from './TextEllipsis'
+export * from './Token'
 export * from './Tooltip'
 export * from './TxInfoBar'
 export * from './Typography'
